Embed all data lists inside curses UI
